+++ tests/testthat/test-make-array.R
@@ -6,7 +6,7 @@ with_mock_HTTP({
         expect_json_equivalent(makeArray(ds[,"gender"], name="Gender array"),
             list(
                 name="Gender array",
-                subvariables=I("/api/datasets/dataset1/variables/gender.json"),
+                subvariables=I("/api/datasets/dataset1/variables/gender/"),
                 type="categorical_array"
             ))
     })
@@ -15,7 +15,7 @@ with_mock_HTTP({
             name="Gender array 2"),
             list(
                 name="Gender array 2",
-                subvariables=I("/api/datasets/dataset1/variables/gender.json"),
+                subvariables=I("/api/datasets/dataset1/variables/gender/"),
                 type="categorical_array"
             ))
     })