+++ tests/testthat/test-progress.R
@@ -4,11 +4,11 @@ with_mock_HTTP({
     test_that("If progress polling gives up, it tells you what to do", {
         with(temp.option(crunch.timeout=0.005), {
             expect_error(
-                expect_output(pollProgress("/api/progress/1.json", wait=0.01),
+                expect_output(pollProgress("/api/progress/1/", wait=0.01),
                     "|================"),
                 paste('Your process is still running on the server. It is',
                     'currently 23% complete. Check',
-                    '`httpcache::uncached(crGET("/api/progress/1.json"))`',
+                    '`httpcache::uncached(crGET("/api/progress/1/"))`',
                     'until it reports 100% complete'),
                 fixed=TRUE)
         })
@@ -17,7 +17,7 @@ with_mock_HTTP({
     ## Setup to test the auto-polling
     fakeProg <- function (progress_url) {
         return(fakeResponse(status_code=202,
-            headers=list(location="/api/datasets.json"),
+            headers=list(location="/api/datasets/"),
             json=list(element="shoji:view", value=progress_url)))
     }
 
@@ -43,7 +43,7 @@ with_mock_HTTP({
             with(temp.option(httpcache.log=logfile), {
                 expect_output(
                     expect_identical(handleAPIresponse(fakeProg("/api/progress/")),
-                        "/api/datasets.json"),
+                        "/api/datasets/"),
                     "=| 100%", fixed=TRUE)
             })
             logs <- loadLogfile(logfile)
@@ -58,7 +58,7 @@ with_mock_HTTP({
                 expect_output(
                     expect_error(
                         expect_message(handleAPIresponse(fakeProg("/api/progress2/")),
-                            "Result URL: /api/datasets.json"),
+                            "Result URL: /api/datasets/"),
                         paste("Education, Commerce, and, uh, oops."), fixed=TRUE),
                     "|  23%", fixed=TRUE)
             })